uoload README.md

+++ src/main/java/university/people/FinanceManager.java
@@ -16,11 +16,13 @@ public class FinanceManager extends User {
 
     public void setRate(int rate) {
         Course.setCreditRate(rate);
+        System.out.println("Rate has been set to " + rate);
     }
 
     public void addScholarship(String scholarshipName, String type, String condition) {
         Scholarship scholarship = ScholarshipFactory.getInstance(scholarshipName, type, condition);
         db.addScholarship(scholarship);
+        System.out.println(scholarshipName + " is added");
     }
 
     public void addTargetConditionToScholarship(String name, String target) {
@@ -30,11 +32,16 @@ public class FinanceManager extends User {
 
     public void deleteScholarship(String name) {
         db.deleteScholarship(name);
+        System.out.println(name + " is deleted.");
     }
 
     public void updateScholarship(String originalName, String newName) {
         Scholarship scholarship = db.findScholarshipByName(originalName);
+        if (scholarship == null) {
+            return;
+        }
         scholarship.setName(newName);
+        System.out.println("Scholarship: " + originalName + " has been renamed as " + newName);
     }
 
     public void browseScholarship(String name) {
@@ -45,13 +52,21 @@ public class FinanceManager extends User {
     public void attachScholarshipToStudent(String studentName, String scholarshipName) {
         Student student = db.findStudentByName(studentName);
         Scholarship scholarship = db.findScholarshipByName(scholarshipName);
+        if (student == null || scholarship == null) {
+            return;
+        }
         db.attachScholarshipToStudent(scholarship, student);
+        System.out.println(studentName + " get " + scholarshipName);
     }
 
     public void deleteScholarshipFromStudent(String studentName, String scholarshipName) {
         Student student = db.findStudentByName(studentName);
         Scholarship scholarship = db.findScholarshipByName(scholarshipName);
+        if (student == null || scholarship == null) {
+            return;
+        }
         db.deleteScholarshipFromStudent(scholarship, student);
+        System.out.println(scholarshipName + " is deleted from " + studentName);
     }
 
     public void browseScholarshipAssignments() {
